Add all-module test, move ccs test
