+++ tests/test_interactive.py
@@ -2,6 +2,7 @@ import multiqc
 
 
 def test_parse_logs_fn_clean_exts(data_dir):
+    multiqc.reset()
     multiqc.parse_logs(
         data_dir / "modules/fastp/SAMPLE.json",
         data_dir / "modules/fastp/single_end",
@@ -12,10 +13,11 @@ def test_parse_logs_fn_clean_exts(data_dir):
 
 
 def test_parse_logs_ignore_samples(data_dir):
+    multiqc.reset()
     multiqc.parse_logs(
         data_dir / "modules/quast/full_metaquast_run",
         ignore_samples=["meta_contigs_2"],
     )
 
-    assert multiqc.list_samples() == ["SRR5442949", "meta_contigs_1", "smalltest"]
-    assert multiqc.list_modules() == ["fastp", "QUAST"]
+    assert multiqc.list_samples() == ["meta_contigs_1"]
+    assert multiqc.list_modules() == ["QUAST"]
